Moved the lexer folder out of analysis, and removed the folder analysis
