+++ pkg/lexer/stateFunctions.go
