+++ pkg/lexer/stateFunctions_test.go
