+++ pkg/lexer/stringMatching.go
