+++ pkg/lexer/token.go
